+++ src/autoComplete/autoCompleteListItem.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import ItemChoice from 'itemChoice'
+import ItemChoice, { ItemChoiceStatus } from 'itemChoice'
 import cc from 'classcat'
 
 export interface AutoCompleteListItemProps {
@@ -8,17 +8,17 @@ export interface AutoCompleteListItemProps {
   readonly children: React.ReactElement<any>,
   readonly className?: Classcat.Class,
   readonly highlighted?: boolean,
-  readonly loading?: boolean,
-  readonly valid?: boolean,
-  readonly validated?: () => void,
+  readonly status?: ItemChoiceStatus,
+  readonly onDoneAnimationEnd?: () => void,
 }
 
 export default class AutoCompleteListItem extends Component <AutoCompleteListItemProps> {
   static defaultProps: Partial<AutoCompleteListItemProps> = {
     highlighted: false,
-    loading: false,
-    valid: false,
+    status: ItemChoice.STATUS.DEFAULT,
   }
+
+  static STATUS = ItemChoiceStatus
 
   onMouseDown = () => {
     this.props.select(this.props.item)
@@ -27,12 +27,11 @@ export default class AutoCompleteListItem extends Component <AutoCompleteListIte
   render() {
     return (
       <ItemChoice
-        loading={this.props.loading}
+        status={this.props.status}
         selected={this.props.highlighted}
         className={cc(['kirk-autoComplete-item', this.props.className])}
         onMouseDown={this.onMouseDown}
-        valid={this.props.valid}
-        validated={this.props.validated}
+        onDoneAnimationEnd={this.props.onDoneAnimationEnd}
       >
         {this.props.children}
       </ItemChoice>