+++ src/button/index.tsx
@@ -9,26 +9,30 @@ import Loader from 'loader'
 import Circle from 'icon/circleIcon'
 import Check from 'icon/checkIcon'
 
+export enum ButtonStatus {
+  PRIMARY = 'primary',
+  SECONDARY = 'secondary',
+  WARNING = 'warning',
+  UNSTYLED = 'unstyled',
+  LOADING = 'loading',
+  CHECKED = 'checked',
+}
+
 export interface ButtonProps {
   readonly type?: string,
   readonly href?: string | JSX.Element,
   readonly children?: string | number | React.ReactNode,
   readonly className?: Classcat.Class,
   readonly title?: string,
-  readonly primary?: boolean,
-  readonly secondary?: boolean,
-  readonly loading?: boolean,
-  readonly valid?: boolean,
-  readonly warning?: boolean,
+  readonly status?: ButtonStatus,
   readonly focus?: boolean,
   readonly icon?: boolean,
-  readonly unstyled?: boolean,
   readonly shadowed?: boolean,
   readonly hidden?: boolean,
   readonly onClick?: (event: React.MouseEvent<HTMLElement>) => void,
   readonly onBlur?: (event: React.FocusEventHandler<HTMLElement>) => void,
   readonly onFocus?: (event: React.FocusEventHandler<HTMLElement>) => void,
-  readonly validated?: (event: Event) => void,
+  readonly onDoneAnimationEnd?: () => void,
   readonly tabIndex?: string,
   readonly disabled?: boolean,
 }
@@ -64,41 +68,27 @@ export const eventHandler = (
 export default class Button extends PureComponent <ButtonProps, ButtonState> {
   private button: HTMLButtonElement
 
+  static STATUS = ButtonStatus
+
   static defaultProps: Partial<ButtonProps> = {
     type: 'button',
     href: '',
     children: '',
     className: '',
-    primary: false,
-    secondary: false,
-    loading: false,
-    valid: false,
-    warning: false,
+    status: ButtonStatus.PRIMARY,
     icon: false,
-    unstyled: false,
     shadowed: false,
     focus: false,
     disabled: false,
   }
 
-  validated = () => {
-    const timeout = parseInt(transition.duration.fast, 10) + transition.callbackDelay
-    setTimeout(this.props.validated, timeout)
-  }
-
   componentDidMount() {
-    if (this.props.valid) {
-      this.validated()
-    }
     if (this.props.focus) {
       this.button.focus()
     }
   }
 
-  componentWillReceiveProps({ valid, focus }: ButtonProps) {
-    if (valid && valid !== this.props.valid) {
-      this.validated()
-    }
+  componentWillReceiveProps({ status, focus }: ButtonProps) {
     if (focus && focus !== this.props.focus) {
       this.button.focus()
     }
@@ -112,9 +102,9 @@ export default class Button extends PureComponent <ButtonProps, ButtonState> {
     const {
       children, className, type, href, title,
       // Modifiers
-      primary, secondary, valid, loading, warning, icon, unstyled, shadowed,
+      status, icon, shadowed,
       // Actions
-      onClick, onBlur, onFocus, validated, focus,
+      onClick, onBlur, onFocus, onDoneAnimationEnd, focus,
       // Extend case of the button for the expand component
       ...attrs,
     } = this.props
@@ -140,20 +130,27 @@ export default class Button extends PureComponent <ButtonProps, ButtonState> {
     typeProps.onFocus = eventHandler(onFocus, typeProps.onFocus)
     typeProps.onBlur = eventHandler(onBlur, typeProps.onBlur)
 
+    const hasLoader = status === ButtonStatus.LOADING || status === ButtonStatus.CHECKED
+    const iconSize = icon || hasLoader
+
     return (
       <Component
         className={cc([
           prefix({ button: true }),
           prefix({
-            primary, secondary, loading, valid, warning, icon, unstyled, shadowed,
+            [status]: status, icon: iconSize, shadowed,
           }, 'kirk-button'),
           className,
         ])}
         {...typeProps}
         {...attrs}
       >
-        {loading && <Loader size={48} inline />}
-        {valid && <Check validate absolute iconColor={color.white} />}
+        {hasLoader && <Loader
+          size={48}
+          inline
+          done={status === ButtonStatus.CHECKED}
+          onDoneAnimationEnd={onDoneAnimationEnd}
+        />}
         <span>{children}</span>
         <style jsx>{style}</style>
       </Component>