+++ src/button/index.unit.tsx
@@ -4,6 +4,8 @@ import Button, { eventHandler } from 'button'
 import ItemChoice from 'itemChoice'
 import CrossIcon from 'icon/crossIcon'
 
+jest.useFakeTimers()
+
 describe('Button', () => {
   class TestLink extends React.PureComponent {
     render() {
@@ -31,8 +33,8 @@ describe('Button', () => {
     const buttonClassName = shallow(<Button className="addClass">blabla</Button>)
     expect(buttonClassName.hasClass('addClass')).toBe(true)
 
-    const button = shallow(<Button primary warning>blabla</Button>)
-    expect(button.hasClass('kirk-button-primary kirk-button-warning')).toBe(true)
+    const button = shallow(<Button status={Button.STATUS.WARNING}>blabla</Button>)
+    expect(button.hasClass('kirk-button-warning')).toBe(true)
   })
 
   it('Should allow for an icon.', () => {
@@ -42,7 +44,7 @@ describe('Button', () => {
   })
 
   it('should allow for a loading state', () => {
-    const button = shallow(<Button loading><CrossIcon /></Button>)
+    const button = shallow(<Button status={Button.STATUS.LOADING}><CrossIcon /></Button>)
     expect(button.hasClass('kirk-button-loading')).toBe(true)
   })
 
@@ -83,16 +85,18 @@ describe('Button', () => {
     expect(secondEvent).toHaveBeenCalledTimes(1)
   })
 
-  jest.useFakeTimers()
-
   it('fires the callback event when valid', () => {
     const event = jest.fn()
-    const button = shallow(<Button validated={event}>blabla</Button>)
-    button.setProps({ valid: true })
-    expect(setTimeout.mock.calls.length).toBe(1)
-    expect(setTimeout.mock.calls[0][0]).toBe(event)
-    expect(setTimeout.mock.calls[0][1]).toBe(500 + 1000) // duration + delay
-    expect(setTimeout).toHaveBeenCalledTimes(1)
+    const button = mount(<Button
+      onDoneAnimationEnd={event}
+      status={Button.STATUS.CHECKED}
+      >
+        blabla
+      </Button>,
+    )
+    expect(event).not.toBeCalled()
+    jest.advanceTimersByTime(1500)
+    expect(event).toBeCalled()
   })
 
   describe('#href', () => {