+++ src/button/style.ts
@@ -106,6 +106,7 @@ export default css`
   :global(.kirk-button-loading) {
     background-color: transparent;
     max-width: initial;
+    border: 0;
   }
 
   :global(.kirk-button-loading:hover),
@@ -129,21 +130,15 @@ export default css`
     margin-left: -24px;
   }
 
-  :global(.kirk-button-valid),
-  :global(.kirk-button-valid:focus),
-  :global(.kirk-button-valid:hover),
-  :global(.kirk-button-valid:active) {
-    background-color: ${color.success};
-    color: ${color.white};
-    max-width: initial;
-    transition: none;
+  :global(.kirk-button-checked) {
+    border: 0;
   }
 
-  :global(.kirk-button-valid[disabled]) {
+  :global(.kirk-button-checked[disabled]) {
     opacity: 1;
   }
 
-  :global(.kirk-button-valid span) {
+  :global(.kirk-button-checked span) {
     opacity: 0;
   }