+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-#Unreleased
-- [...]
+# Unreleased
+- [Loader] Now handles the "green check" animation and callback
+- [Button & ItemChoice] API breaking changes, now use a unique `status` instead of multiple props to avoid overlaps of properties and styles
+- [Radio & AutoComplete] Repercussions from ItemChoice API changes
 
 #v0.1.1
 - NEW UneditableTextField component